client/requirements.txt
rel==0.4.7
websocket-client==1.4.0
